Client_server_app_example/requirements.txt
aniso8601==9.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scikit-learn==0.24.2
scipy==1.9.1
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
